pyproject.toml
[tool.poetry]
name = "frontiers_yildizetal"
version = "0.1.0"
description = "Python package used in Yildiz et al. (2022) in Frontiers"
authors = ["Anil Yildiz"]
license = "GNU General Public License v3.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.10"
rasterio = "^1.3.2"
numpy = "^1.23.2"
pandas = "^1.4.4"
PyYAML = "^6.0"
setuptools = "^65.3.0"
scipy = "^1.9.1"
scikit-learn = "^1.1.2"
rpy2 = "^3.5.4"
ipykernel = "^6.15.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
